Create main.yml
Started customizing the template actions file Github provides. I need to figure out how to push the code to the production server using the SSH key provided via a repo secret, which turned out to be a bit of a hassle. I do not have time to fix this right now, but I will get back to this as soon as I can.